analyzer/requirements.txt
pika==1.3.2
requests==2.31.0
influxdb==5.3.2
tenacity==8.2.3
pyyaml==6.0.1
